requirements.txt
bcrypt==4.0.1
black==23.9.1
blinker==1.6.2
click==8.1.7
dnspython==2.4.2
email-validator==2.0.0.post2
flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
flask-sqlalchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
login==0.0.6
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.1
platformdirs==3.10.0
psycopg2-binary==2.9.7
SQLAlchemy==2.0.20
tomli==2.0.1
typing-extensions==4.7.1
werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
